chore: Do not bundle tinted with this library
This drastically decreases the size of the build for this library.  We
can rely on our users to decide how to bundle their application, because
we have properly declared tined as a dependency of this library.

This commit reduces the size of the bundle from 2080 lines to 208 lines.
That is a perfect TEN TIMES reduction.